chart: parse charts with negative signed axis IDs
Battle-testing the charts reader against ~187 chart-bearing real-world
Common Crawl OOXML packages surfaced a common hard failure: encoding/xml
rejected an entire chart part with

  strconv.ParseUint: parsing "-2042759216": invalid syntax

whenever a c:axId / c:crossAx carried a negative value. Although the
schema types these as xsd:unsignedInt, Microsoft PowerPoint and Excel
routinely emit axis identifiers as negative signed int32 values, so a
uint32 model made Charts() silently drop every such chart across several
corpus files.

Widen CT_UnsignedInt.Val from uint32 to int64 (as Apache POI models axId
as long). int64 accepts both the full unsignedInt range and the de-facto
signed ids, and serializes every value to identical text, so new-chart
round-trip fidelity is unchanged; opened chart parts are byte-preserved
regardless. Add a regression test parsing a chartSpace whose axId/crossAx
are negative.

Corpus result after the fix: 185/187 chart-bearing packages pass
open -> Charts() -> zero-mod save -> reopen -> byte fidelity; 1644 charts
read via the public API; 0 chart-part fidelity failures and 0 panics. The
two non-passing files are a whitespace-preserving producer (pre-existing
wontfix) and a corrupt source zip, neither chart-related. Remaining
unparsed parts are all unsupported chart types (3D/doughnut/radar/surface/
ofPie), which degrade gracefully.

+++ chart/serialize.go
@@ -351,7 +351,7 @@ func axIDs(ids ...uint32) []*dmlchart.UnsignedInt {
 
 func boolElem(v bool) *dmlchart.Boolean { return &dmlchart.Boolean{Val: v} }
 func uintElem(v uint32) *dmlchart.UnsignedInt {
-	return &dmlchart.UnsignedInt{Val: v}
+	return &dmlchart.UnsignedInt{Val: int64(v)}
 }
 func boolPtr(v bool) *bool { return &v }
